requirements.txt
fastapi 
python-dotenv
uvicorn
google-generativeai
pandas
